requirements.txt
# Core dependencies for Exam-Helper application

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-mistralai>=0.0.4
langchain-google-genai>=0.0.6
langchain-pinecone>=0.0.3
langchain-groq>=0.0.3

# Vector database and embeddings
pinecone-client>=3.0.0

# Web search and data processing
duckduckgo-search>=4.0.0
requests>=2.31.0

# Streamlit and UI components
streamlit>=1.28.0
st-multimodal-chatinput>=0.1.0
streamlit-carousel>=0.1.0

# Image processing
Pillow>=10.0.0

# Google AI
google-generativeai>=0.3.0

# Configuration management
PyYAML>=6.0

# Data validation
pydantic>=2.0.0

# Additional utilities
python-dotenv>=1.0.0